Move SymPyDeprecationWarning into sympy.utilities.exceptions.
