+++ sympy/matrices/matrices.py
@@ -17,7 +17,7 @@ from sympy.printing import sstr
 from sympy.functions.elementary.trigonometric import cos, sin
 
 from sympy.core.compatibility import callable, reduce
-from sympy.core.exceptions import SymPyDeprecationWarning
+from sympy.utilities.exceptions import SymPyDeprecationWarning
 from sympy.core.decorators import call_highest_priority
 
 import random
